RUN: Use toolchain to patch command line in `createGeneralCommandLine`
