+++ src/main/kotlin/org/rust/cargo/toolchain/RsToolchain.kt
@@ -5,13 +5,21 @@
 
 package org.rust.cargo.toolchain
 
+import com.intellij.execution.configuration.EnvironmentVariablesData
 import com.intellij.execution.configurations.GeneralCommandLine
+import com.intellij.execution.configurations.PtyCommandLine
 import com.intellij.execution.process.ProcessHandler
 import com.intellij.openapi.extensions.ExtensionPointName
+import com.intellij.openapi.util.SystemInfo
 import com.intellij.util.io.exists
+import com.intellij.util.net.HttpConfigurable
 import com.intellij.util.text.SemVer
+import org.rust.cargo.CargoConstants
 import org.rust.cargo.toolchain.flavors.RsToolchainFlavor
 import org.rust.cargo.toolchain.tools.Cargo
+import org.rust.openapiext.GeneralCommandLine
+import org.rust.openapiext.withWorkDirectory
+import java.io.File
 import java.nio.file.Path
 import java.nio.file.Paths
 
@@ -34,6 +42,11 @@ abstract class RsToolchain(val location: Path) {
     abstract val fileSeparator: String
 
     fun looksLikeValidToolchain(): Boolean = RsToolchainFlavor.getFlavor(location) != null
+
+    /**
+     * Patches passed command line to make it runnable in remote context.
+     */
+    abstract fun <T : GeneralCommandLine> patchCommandLine(commandLine: T): T
 
     abstract fun startProcess(commandLine: GeneralCommandLine): ProcessHandler
 
@@ -74,6 +87,48 @@ abstract class RsToolchain(val location: Path) {
     }
 
     override fun hashCode(): Int = location.hashCode()
+
+    fun createGeneralCommandLine(
+        executable: Path,
+        workingDirectory: Path,
+        redirectInputFrom: File?,
+        backtraceMode: BacktraceMode,
+        environmentVariables: EnvironmentVariablesData,
+        parameters: List<String>,
+        emulateTerminal: Boolean,
+        patchToRemote: Boolean = true,
+        http: HttpConfigurable = HttpConfigurable.getInstance()
+    ): GeneralCommandLine {
+        var commandLine = GeneralCommandLine(executable)
+            .withWorkDirectory(workingDirectory)
+            .withInput(redirectInputFrom)
+            .withEnvironment("TERM", "ansi")
+            .withParameters(parameters)
+            .withCharset(Charsets.UTF_8)
+            .withRedirectErrorStream(true)
+        withProxyIfNeeded(commandLine, http)
+
+        when (backtraceMode) {
+            BacktraceMode.SHORT -> commandLine.withEnvironment(CargoConstants.RUST_BACKTRACE_ENV_VAR, "short")
+            BacktraceMode.FULL -> commandLine.withEnvironment(CargoConstants.RUST_BACKTRACE_ENV_VAR, "full")
+            BacktraceMode.NO -> Unit
+        }
+
+        environmentVariables.configureCommandLine(commandLine, true)
+
+        if (emulateTerminal) {
+            if (!SystemInfo.isWindows) {
+                commandLine.environment["TERM"] = "xterm-256color"
+            }
+            commandLine = PtyCommandLine(commandLine).withInitialColumns(PtyCommandLine.MAX_COLUMNS)
+        }
+
+        if (patchToRemote) {
+            commandLine = patchCommandLine(commandLine)
+        }
+
+        return commandLine
+    }
 
     companion object {
         val MIN_SUPPORTED_TOOLCHAIN = SemVer.parseFromText("1.32.0")!!
